Changed it such that a new pope and emperor gets elected every 30 years
Changed it such that a new pope and emperor gets elected every 30 years - this prevents the imperial offices from remaining empty. This will still make the bug occur, but as it will happen in the same year (every 30 years) for both the emperor and the pope the disruption should be much less than normal.